将session里面的uid类型换成了Long,遂修改一些页面

+++ src/servlet/admin/changePassword.java
@@ -23,10 +23,10 @@ public class changePassword extends HttpServlet {
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HttpSession session = request.getSession();
         PrintWriter out = response.getWriter();
-        Integer uid = (Integer)session.getAttribute("uid");
+        Long uid = (Long)session.getAttribute("uid");
         String username = (String)session.getAttribute("username");
 
-        User user = UserDao.getUserByUidOrUsername(uid,username);
+        User user = UserDao.getUserByUidOrUsername(uid.intValue(),username);
         String oldpw = request.getParameter("oldpw");
         String newpw = request.getParameter("newpw");
         if(oldpw==null || newpw==null){
@@ -34,7 +34,7 @@ public class changePassword extends HttpServlet {
             return;
         }
         if(user!=null && user.getPassword().equals(oldpw)){
-            if(UserDao.updatePassword(uid,newpw))
+            if(UserDao.updatePassword(uid.intValue(),newpw))
                 NetUtils.writeResultToBrowser(out,true,"修改密码成功!");
             else
                 NetUtils.writeResultToBrowser(out,false,"修改密码失败,数据库错误或无此人信息!");
